Move toolkit details components into toolkit/ subfolder
